+++ public/index.php
@@ -2,4 +2,13 @@
 
 declare(strict_types=1);
 
-echo "index.php";
+require_once dirname(__DIR__) . '/vendor/autoload.php';
+
+use Quinen\Cms\Http\Request;
+
+echo '<pre><code>' . var_export([
+        Request::getAsString('toto'),
+        Request::getAsInt('toto'),
+        Request::getAsInt('titi'),
+        $_SERVER
+    ], true) . '</code></pre>';