+++ src/Dispatcher.php
@@ -1,0 +1,10 @@
+<?php
+
+
+namespace Quinen\Cms;
+
+
+class Dispatcher
+{
+
+}